+++ backend/src/main/java/com/example/backend/service/quiz/QuizService.java
@@ -133,7 +133,7 @@ public class QuizService {
         return quizDTOS;
     }
 
-    public List<QuizDTO>getAllQuizzesByLevelandLanguageId(long level, long languageId) {
+    public List<QuizDTO>getAllQuizzesByLevelandLanguageId(int level, int languageId) {
         List<QuizDTO> quizDTOS = new ArrayList<>();
         quizRepository.getAllByLevelAndLanguageId(level, languageId).forEach(quiz -> quizDTOS.add(quizDTOConverterService.apply(quiz, null)));
         return quizDTOS;
